# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "daikon-studio",
    "version": "0.0.0+dev"
  },
  "paths": {
...
    "/api/v1/chemcellar/protocols/{protocol_id}/runs": {
      "get": {
        "tags": [
          "chemcellar"
        ],
        "summary": "List Chemcellar Runs",
        "operationId": "list_chemcellar_runs_api_v1_chemcellar_protocols__protocol_id__runs_get",
        "parameters": [
          {
            "name": "protocol_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Protocol Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChemCellarRunResponse"
                  },
                  "title": "Response List Chemcellar Runs Api V1 Chemcellar Protocols  Protocol Id  Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chemcellar/runs/{run_id}/import": {
      "post": {
        "tags": [
          "chemcellar"
        ],
        "summary": "Import Chemcellar Run",
        "operationId": "import_chemcellar_run_api_v1_chemcellar_runs__run_id__import_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChemCellarImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/protocols": {
      "post": {
        "tags": [
          "protocols"
        ],
        "summary": "Train Protocol",
        "operationId": "train_protocol_api_v1_protocols_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainProtocolBody"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "protocols"
        ],
        "summary": "List Protocols",
        "operationId": "list_protocols_api_v1_protocols_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "dataset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ProtocolResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChemCellarImportResponse": {
        "properties": {
          "upload_ref": {
            "type": "string",
            "format": "uuid",
            "title": "Upload Ref"
          },
          "compound_count": {
            "type": "integer",
            "title": "Compound Count"
          },
          "without_structure": {
            "type": "integer",
            "title": "Without Structure"
          },
          "sample": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sample"
          },
          "source": {
            "$ref": "#/components/schemas/RunSourceResponse"
          }
        },
        "type": "object",
        "required": [
          "upload_ref",
          "compound_count",
          "without_structure",
          "sample",
          "source"
        ],
        "title": "ChemCellarImportResponse"
      },
...
      "ChemCellarRunResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "protocol_id": {
            "type": "string",
            "format": "uuid",
            "title": "Protocol Id"
          },
          "run_date": {
            "type": "string",
            "format": "date",
            "title": "Run Date"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "measured_count": {
            "type": "integer",
            "title": "Measured Count"
          },
          "plate_count": {
            "type": "integer",
            "title": "Plate Count"
          },
          "plate_barcodes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Plate Barcodes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "protocol_id",
          "run_date",
          "status",
          "measured_count",
          "plate_count",
          "plate_barcodes"
        ],
        "title": "ChemCellarRunResponse"
      },
...
      "Direction": {
        "type": "string",
        "enum": [
          "high",
          "low"
        ],
        "title": "Direction"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_ProtocolResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProtocolResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PaginatedResponse[ProtocolResponse]"
      },
...
      "ProtocolResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "engine_id": {
            "type": "string",
            "title": "Engine Id"
          },
          "artifact_uri": {
            "type": "string",
            "title": "Artifact Uri"
          },
          "readouts": {
            "items": {
              "$ref": "#/components/schemas/ReadoutResponse"
            },
            "type": "array",
            "title": "Readouts"
          },
          "conditions": {
            "additionalProperties": true,
            "type": "object",
            "title": "Conditions"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "is_locked": {
            "type": "boolean",
            "title": "Is Locked"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "parent_protocol_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Protocol Id"
          },
          "protocol_version": {
            "type": "integer",
            "title": "Protocol Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "can_delete": {
            "type": "boolean",
            "title": "Can Delete"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "name",
          "dataset_id",
          "engine_id",
          "artifact_uri",
          "readouts",
          "conditions",
          "status",
          "is_locked",
          "published_at",
          "parent_protocol_id",
          "protocol_version",
          "created_at",
          "can_delete"
        ],
        "title": "ProtocolResponse"
      },
...
      "ReadoutResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/ReadoutType"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "direction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Direction"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type",
          "unit",
          "direction",
          "description",
          "threshold"
        ],
        "title": "ReadoutResponse",
        "description": "One declared output of a trained Protocol.\n\nTyped rather than a bare `dict` because `unit` and `direction` are the\nwhole point of a Readout: they are what let a predicted IC50 be lined up\nagainst a measured one instead of being a bare float. That pairing was\ndropped at four separate boundaries during the backend build; handing the\nclient `dict[str, Any]` here would have invited a fifth.\n\n`direction` is narrowed to the `Direction` enum even though `Readout` types\nit as `str`. Every value `derive_readouts` can produce is a `Direction`\nmember or `None`, so the narrower contract is honest and gives a client a\nclosed union instead of an open string."
      },
      "ReadoutType": {
        "type": "string",
        "enum": [
          "numeric",
          "probability",
          "class"
        ],
        "title": "ReadoutType"
      },
...
      "RunResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "result_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Uri"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "protocol_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol Id"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "lane": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lane"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "requested_by": {
            "type": "string",
            "format": "uuid",
            "title": "Requested By"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunSourceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "kind",
          "status",
          "progress",
          "phase",
          "result_uri",
          "error_message",
          "protocol_id",
          "metrics",
          "lane",
          "name",
          "requested_by",
          "source",
          "created_at"
        ],
        "title": "RunResponse"
      },
      "RunSourceResponse": {
        "properties": {
          "app": {
            "type": "string",
            "const": "chemcellar",
            "title": "App"
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "protocol_id": {
            "type": "string",
            "format": "uuid",
            "title": "Protocol Id"
          },
          "protocol_name": {
            "type": "string",
            "title": "Protocol Name"
          },
          "run_date": {
            "type": "string",
            "format": "date",
            "title": "Run Date"
          },
          "compounds_without_structure": {
            "type": "integer",
            "title": "Compounds Without Structure",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "app",
          "run_id",
          "protocol_id",
          "protocol_name",
          "run_date"
        ],
        "title": "RunSourceResponse",
        "description": "Where a prediction run's compounds came from, when not from an uploaded file."
      },
...
      "TrainProtocolBody": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 256,
            "title": "Name"
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "engine_id": {
            "type": "string",
            "title": "Engine Id"
          },
          "conditions": {
            "additionalProperties": true,
            "type": "object",
            "title": "Conditions"
          },
          "baseline_engine_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Engine Id"
          },
          "baseline_conditions": {
            "additionalProperties": true,
            "type": "object",
            "title": "Baseline Conditions"
          },
          "tune_cutoffs": {
            "type": "boolean",
            "title": "Tune Cutoffs",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "dataset_id",
          "engine_id",
          "conditions"
        ],
        "title": "TrainProtocolBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}